requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
Werkzeug==3.0.1
sendgrid==6.11.0
